requirements.txt
# ===== Core DL / 3D / pipeline =====
torch==2.3.0+cu118
torchvision==0.18.0+cu118
xformers==0.0.33.post1
accelerate==1.11.0
diffusers==0.20.2
transformers==4.29.2
huggingface_hub==0.25.2
einops==0.8.1
pytorch_msssim==1.0.0
git+https://github.com/facebookresearch/segment-anything.git

# ===== Numeric / Sci =====
numpy==2.3.5
scipy==1.16.3
h5py==3.15.1
quaternion==3.5.2.post4
scikit_learn==1.7.2

# ===== 3D / geometry =====
open3d==0.19.0
trimesh==4.9.0
plyfile==1.1.3

# ===== Image / IO / plotting =====
imageio==2.37.2
Pillow==12.0.0
opencv_python_headless==4.12.0.88
matplotlib==3.10.7

# ===== Config / utils =====
omegaconf==2.3.0
PyYAML==6.0.3
tqdm==4.67.1
fvcore==0.1.5.post20221221
iopath==0.1.10
roma==1.5.4
lightning==2.3.3
onnxruntime==1.23.2
